efm32: Started pre-fetching OUT and SETUP packets.
This seems to allow things to work where they haven't before. Right now it can
get past the set address packet and fails on an unrecognized packet type later
on.

It currently requires a modification to the usb-device library to get that far
where it ignores OUT requests while a control pipe is in the IDLE state.