efm32: Switched polling from GRXSTSP to GRXSTSR.
This follows the original/generic implementation and fixes a possible limitation
in this one where it could discard a packet that it wasn't yet able to place
into the RX queue.

Now, the current packet info is popped off the queue only when the poll method
should read or discard the data.